app/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
decorator==5.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
gunicorn==22.0.0
idna==3.7
ipinfo==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
maxminddb==2.6.2
multidict==6.0.5
packaging==24.1
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
ratelim==0.1.6
requests==2.32.3
setuptools==73.0.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
